krishi-ai-backend/requirements.txt
# Krishi AI Backend Requirements

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Google Gemini SDK for TTS
google-genai>=0.3.0
# Alternative: google-generativeai>=0.3.0

# Environment and Configuration
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.25.0

# For async support
anyio>=3.7.0
